moved event to components to separate folder
