+++ portal/src/components/App.tsx
@@ -2,7 +2,7 @@ import Login from "./Login";
 import { useActions } from "../hooks/useActions";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import { useTypedSelector } from "../hooks/useTypedSelector";
-import EditEvent from "./EditEvent";
+import EditEvent from "./Events/EditEvent";
 import Landing from "./Landing";
 const App = () => {
   const { data, error, loading } = useTypedSelector((state) => state.auth);
